Merge NuttX 10.1.0

+++ arch/arm/src/max326xx/max32660/max32660_spim.c
@@ -1,35 +1,20 @@
 /****************************************************************************
- * arch/arm/src/max326/max326_spi.c
+ * arch/arm/src/max326xx/max32660/max32660_spim.c
  *
- *   Copyright (C) 2018 Gregory Nutt. All rights reserved.
- *   Author: Gregory Nutt <gnutt@nuttx.org>
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.  The
+ * ASF licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the
+ * License.  You may obtain a copy of the License at
  *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in
- *    the documentation and/or other materials provided with the
- *    distribution.
- * 3. Neither the name NuttX nor the names of its contributors may be
- *    used to endorse or promote products derived from this software
- *    without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
- * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
- * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
  *
  ****************************************************************************/
 
@@ -74,8 +59,8 @@
 #include <nuttx/semaphore.h>
 #include <nuttx/spi/spi.h>
 
-#include "up_internal.h"
-#include "up_arch.h"
+#include "arm_internal.h"
+#include "arm_arch.h"
 
 #include "chip.h"
 #include "hardware/max326_pinmux.h"
@@ -193,7 +178,7 @@ static void spi_setbits(struct spi_dev_s *dev, int nbits);
 static int  spi_hwfeatures(struct spi_dev_s *dev,
               spi_hwfeatures_t features);
 #endif
-static uint16_t spi_send(struct spi_dev_s *dev, uint16_t wd);
+static uint32_t spi_send(struct spi_dev_s *dev, uint32_t wd);
 static void spi_exchange(struct spi_dev_s *dev, const void *txbuffer,
               void *rxbuffer, size_t nwords);
 #ifndef CONFIG_SPI_EXCHANGE
@@ -247,7 +232,10 @@ static const struct spi_ops_s g_sp0iops =
 
 static struct max326_spidev_s g_spi0dev =
 {
-  .dev      = { &g_sp0iops },
+  .dev      =
+    {
+      &g_sp0iops
+    },
   .base     = MAX326_SPI0_BASE,
 #ifdef CONFIG_MAX326_SPI_INTERRUPTS
   .irq      = MAX326_IRQ_SPI,
@@ -624,7 +612,8 @@ static int spi_poll(struct max326_spidev_s *priv)
   /* Break out if we've transmitted all the bytes and not receiving */
 
   if ((priv->rxbuffer == NULL) && priv->txbytes == length &&
-      (spi_getreg(priv, MAX326_SPI_DMA_OFFSET) & SPI_DMA_TXFIFOCNT_MASK) == 0)
+      (spi_getreg(priv, MAX326_SPI_DMA_OFFSET) &
+       SPI_DMA_TXFIFOCNT_MASK) == 0)
     {
       goto done;
     }
@@ -704,7 +693,9 @@ static int spi_poll(struct max326_spidev_s *priv)
           inten |= SPI_INT_RXLEVEL;
         }
 
-      /* Break out if we've received all the bytes and we're not transmitting */
+      /* Break out if we've received all the bytes and we're not
+       * transmitting.
+       */
 
       if (priv->txbuffer == NULL && priv->rxbytes == length)
         {
@@ -712,7 +703,7 @@ static int spi_poll(struct max326_spidev_s *priv)
         }
     }
 
-  /* Break out once we've transmitted and received all of the data */
+  /* Break out once we've transmitted and received all of the data. */
 
   if (priv->rxbytes == length && priv->txbytes == length)
     {
@@ -790,8 +781,10 @@ static int spi_interrupt(int irq, void *context, void *arg)
           /* Check if there is more Rx data to be read */
 
           regval  = spi_getreg(priv, MAX326_SPI_DMA_OFFSET);
-          rxavail = (regval & SPI_DMA_TXFIFOCNT_MASK) >> SPI_DMA_TXFIFOCNT_SHIFT;
-          rxlevel = (regval & SPI_DMA_RXFIFOLVL_MASK) >> SPI_DMA_RXFIFOLVL_SHIFT;
+          rxavail = (regval & SPI_DMA_TXFIFOCNT_MASK) >>
+                     SPI_DMA_TXFIFOCNT_SHIFT;
+          rxlevel = (regval & SPI_DMA_RXFIFOLVL_MASK) >>
+                     SPI_DMA_RXFIFOLVL_SHIFT;
         }
       while (/* RX buffer != NULL && */ rxavail > rxlevel);
     }
@@ -804,12 +797,12 @@ static int spi_interrupt(int irq, void *context, void *arg)
  * Name: spi_lock
  *
  * Description:
- *   On SPI busses where there are multiple devices, it will be necessary to
- *   lock SPI to have exclusive access to the busses for a sequence of
+ *   On SPI buses where there are multiple devices, it will be necessary to
+ *   lock SPI to have exclusive access to the buses for a sequence of
  *   transfers.  The bus should be locked before the chip is selected. After
  *   locking the SPI bus, the caller should then also call the setfrequency,
  *   setbits, and setmode methods to make sure that the SPI is properly
- *   configured for the device.  If the SPI buss is being shared, then it
+ *   configured for the device.  If the SPI bus is being shared, then it
  *   may have been left in an incompatible state.
  *
  * Input Parameters:
@@ -867,7 +860,7 @@ static void spi0_select(struct spi_dev_s *dev, uint32_t devid, bool selected)
   DEBUGASSERT(dev != NULL);
   max326_spi0select(dev, devid, selected);
 
-  /* The hardware requires that the SPI block be disabled and renabled at
+  /* The hardware requires that the SPI block be disabled and re-enabled at
    * end of each transaction to cancel the on ongoing transaction (while
    * chip select is inactive).
    */
@@ -998,12 +991,14 @@ static uint32_t spi_setfrequency(struct spi_dev_s *dev, uint32_t frequency)
         }
 
       regval  = spi_getreg(priv, MAX326_SPI_CLKCFG_OFFSET);
-      regval &= ~(SPI_CLKCFG_LO_MASK | SPI_CLKCFG_HI_MASK | SPI_CLKCFG_SCALE_MASK);
-      regval |= (SPI_CLKCFG_LO(high) | SPI_CLKCFG_HI(high) | SPI_CLKCFG_SCALE(scale));
+      regval &= ~(SPI_CLKCFG_LO_MASK | SPI_CLKCFG_HI_MASK |
+                  SPI_CLKCFG_SCALE_MASK);
+      regval |= (SPI_CLKCFG_LO(high) | SPI_CLKCFG_HI(high) |
+                 SPI_CLKCFG_SCALE(scale));
       spi_putreg(priv, MAX326_SPI_CLKCFG_OFFSET, regval);
 
-      /* Save the frequency selection so that subsequent reconfigurations will be
-       * faster.
+      /* Save the frequency selection so that subsequent reconfigurations
+       * will be faster.
        */
 
       spiinfo("Frequency %d->%d\n", frequency, actual);
@@ -1072,7 +1067,9 @@ static void spi_setmode(struct spi_dev_s *dev, enum spi_mode_e mode)
 
         spi_modify_ctrl2(priv, setbits, clrbits);
 
-        /* Save the mode so that subsequent re-configurations will be faster */
+        /* Save the mode so that subsequent re-configurations will be
+         * faster.
+         */
 
         priv->mode = mode;
     }
@@ -1113,8 +1110,8 @@ static void spi_setbits(struct spi_dev_s *dev, int nbits)
 
       priv->data16 = (nbits > 8);
 
-      /* Save the selection so the subsequence re-configurations will be
-       * faster
+      /* Save the selection so that subsequent re-configurations will be
+       * faster.
        */
 
       priv->nbits  = nbits;
@@ -1160,9 +1157,9 @@ static int spi_hwfeatures(struct spi_dev_s *dev, spi_hwfeatures_t features)
  *
  ****************************************************************************/
 
-static uint16_t spi_send(struct spi_dev_s *dev, uint16_t wd)
+static uint32_t spi_send(struct spi_dev_s *dev, uint32_t wd)
 {
-  uint16_t ret;
+  uint32_t ret;
 
   spiinfo("wd=%04u\n", wd);
   spi_exchange(dev, &wd, &ret, 1);
@@ -1190,8 +1187,8 @@ static uint16_t spi_send(struct spi_dev_s *dev, uint16_t wd)
  *
  ****************************************************************************/
 
-static void spi_exchange(struct spi_dev_s *dev, const void *txbuffer, void *rxbuffer,
-                         size_t nwords)
+static void spi_exchange(struct spi_dev_s *dev, const void *txbuffer,
+                         void *rxbuffer, size_t nwords)
 {
   struct max326_spidev_s *priv = (struct max326_spidev_s *)dev;
 #ifndef CONFIG_MAX326_SPI_INTERRUPTS
@@ -1331,11 +1328,11 @@ static void spi_exchange(struct spi_dev_s *dev, const void *txbuffer, void *rxbu
  * Input Parameters:
  *   dev      - Device-specific state data
  *   txbuffer - A pointer to the buffer of data to be sent
- *   nwords   - the length of data to send from the buffer in number of words.
- *              The wordsize is determined by the number of bits-per-word
- *              selected for the SPI interface.  If nbits <= 8, the data is
- *              packed into uint8_t's; if nbits >8, the data is packed into
- *              uint16_t's
+ *   nwords   - the length of data to send from the buffer in number of
+ *              words.  The wordsize is determined by the number of
+ *              bits-per-word selected for the SPI interface.  If nbits <= 8,
+ *              the data is packed into uint8_t's; if nbits >8, the data is
+ *              packed into uint16_t's
  *
  * Returned Value:
  *   None
@@ -1343,7 +1340,8 @@ static void spi_exchange(struct spi_dev_s *dev, const void *txbuffer, void *rxbu
  ****************************************************************************/
 
 #ifndef CONFIG_SPI_EXCHANGE
-static void spi_sndblock(struct spi_dev_s *dev, const void *txbuffer, size_t nwords)
+static void spi_sndblock(struct spi_dev_s *dev, const void *txbuffer,
+                         size_t nwords)
 {
   spiinfo("txbuffer=%p nwords=%d\n", txbuffer, nwords);
   spi_exchange(dev, txbuffer, NULL, nwords);
@@ -1359,11 +1357,11 @@ static void spi_sndblock(struct spi_dev_s *dev, const void *txbuffer, size_t nwo
  * Input Parameters:
  *   dev      - Device-specific state data
  *   rxbuffer - A pointer to the buffer in which to receive data
- *   nwords   - the length of data that can be received in the buffer in number
- *              of words.  The wordsize is determined by the number of bits-per-word
- *              selected for the SPI interface.  If nbits <= 8, the data is
- *              packed into uint8_t's; if nbits >8, the data is packed into
- *              uint16_t's
+ *   nwords   - the length of data that can be received in the buffer in
+ *              number of words.  The wordsize is determined by the number of
+ *              bits-per-word selected for the SPI interface.  If nbits <= 8,
+ *              the data is packed into uint8_t's; if nbits >8, the data is
+ *              packed into uint16_t's
  *
  * Returned Value:
  *   None
@@ -1371,7 +1369,8 @@ static void spi_sndblock(struct spi_dev_s *dev, const void *txbuffer, size_t nwo
  ****************************************************************************/
 
 #ifndef CONFIG_SPI_EXCHANGE
-static void spi_recvblock(struct spi_dev_s *dev, void *rxbuffer, size_t nwords)
+static void spi_recvblock(struct spi_dev_s *dev, void *rxbuffer,
+                          size_t nwords)
 {
   spiinfo("rxbuffer=%p nwords=%d\n", rxbuffer, nwords);
   spi_exchange(dev, NULL, rxbuffer, nwords);
@@ -1382,8 +1381,8 @@ static void spi_recvblock(struct spi_dev_s *dev, void *rxbuffer, size_t nwords)
  * Name: spi_bus_initialize
  *
  * Description:
- *   Initialize the selected SPI bus in its default state (Master, 8-bit, mode 0,
- *   etc.)
+ *   Initialize the selected SPI bus in its default state (Master, 8-bit,
+ *   mode 0, etc.)
  *
  * Input Parameters:
  *   priv   - private SPI device structure
@@ -1405,7 +1404,8 @@ static void spi_bus_initialize(struct max326_spidev_s *priv)
 
   /* Setup slaved select timing (even in Master mode?) */
 
-  regval = (SPI_SSTIME_SSACT1(1) | SPI_SSTIME_SSACT2(1) | SPI_SSTIME_SSINACT(1));
+  regval = (SPI_SSTIME_SSACT1(1) | SPI_SSTIME_SSACT2(1) |
+            SPI_SSTIME_SSINACT(1));
   spi_putreg(priv, MAX326_SPI_SSTIME_OFFSET, regval);
 
   /* Configure CTRL0. Default configuration:
