+++ fs/nxffs/nxffs_reformat.c
@@ -1,37 +1,20 @@
 /****************************************************************************
  * fs/nxffs/nxffs_reformat.c
  *
- *   Copyright (C) 2011, 2013 Gregory Nutt. All rights reserved.
- *   Author: Gregory Nutt <gnutt@nuttx.org>
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.  The
+ * ASF licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the
+ * License.  You may obtain a copy of the License at
  *
- * References: Linux/Documentation/filesystems/romfs.txt
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in
- *    the documentation and/or other materials provided with the
- *    distribution.
- * 3. Neither the name NuttX nor the names of its contributors may be
- *    used to endorse or promote products derived from this software
- *    without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
- * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
- * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
- * OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED
- * AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
- * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
- * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
  *
  ****************************************************************************/
 
@@ -41,6 +24,7 @@
 
 #include <nuttx/config.h>
 
+#include <stdint.h>
 #include <string.h>
 #include <errno.h>
 #include <debug.h>
@@ -96,7 +80,7 @@ static int nxffs_format(FAR struct nxffs_volume_s *volume)
       ret = MTD_ERASE(volume->mtd, eblock, 1);
       if (ret < 0)
         {
-          ferr("ERROR: Erase block %d failed: %d\n", eblock, ret);
+          ferr("ERROR: Erase block %jd failed: %d\n", (intmax_t)eblock, ret);
           return ret;
         }
 
@@ -106,7 +90,8 @@ static int nxffs_format(FAR struct nxffs_volume_s *volume)
       nxfrd = MTD_BWRITE(volume->mtd, lblock, volume->blkper, volume->pack);
       if (nxfrd != volume->blkper)
         {
-          ferr("ERROR: Write erase block %d failed: %d\n", lblock, nxfrd);
+          ferr("ERROR: Write erase block %jd failed: %zd\n",
+               (intmax_t)lblock, nxfrd);
           return -EIO;
         }
     }
@@ -157,7 +142,8 @@ static int nxffs_badblocks(FAR struct nxffs_volume_s *volume)
       nxfrd  = MTD_BREAD(volume->mtd, lblock, volume->blkper, volume->pack);
       if (nxfrd != volume->blkper)
         {
-          ferr("ERROR: Read erase block %d failed: %d\n", lblock, nxfrd);
+          ferr("ERROR: Read erase block %jd failed: %zd\n",
+               (intmax_t)lblock, nxfrd);
           return -EIO;
         }
 #endif
@@ -178,7 +164,8 @@ static int nxffs_badblocks(FAR struct nxffs_volume_s *volume)
            i++, block++, blkptr += volume->geo.blocksize)
 #endif
         {
-          FAR struct nxffs_block_s *blkhdr = (FAR struct nxffs_block_s *)blkptr;
+          FAR struct nxffs_block_s *blkhdr =
+            (FAR struct nxffs_block_s *)blkptr;
 
           /* Assume that this is a good block until we learn otherwise */
 
@@ -207,8 +194,8 @@ static int nxffs_badblocks(FAR struct nxffs_volume_s *volume)
           if (memcmp(blkhdr->magic, g_blockmagic, NXFFS_MAGICSIZE) != 0 ||
               blkhdr->state != BLOCK_STATE_GOOD)
             {
-              /* The block is not formated with the NXFFS magic bytes or else
-               * the block is specifically marked bad.
+              /* The block is not formatted with the NXFFS magic bytes or
+               * else the block is specifically marked bad.
                */
 
               good = false;
@@ -220,9 +207,11 @@ static int nxffs_badblocks(FAR struct nxffs_volume_s *volume)
 
           else
             {
-              size_t blocksize = volume->geo.blocksize - SIZEOF_NXFFS_BLOCK_HDR;
-              size_t erasesize = nxffs_erased(&blkptr[SIZEOF_NXFFS_BLOCK_HDR],
-                                              blocksize);
+              size_t blocksize = volume->geo.blocksize -
+                                 SIZEOF_NXFFS_BLOCK_HDR;
+              size_t erasesize = nxffs_erased(
+                                     &blkptr[SIZEOF_NXFFS_BLOCK_HDR],
+                                     blocksize);
               good = (blocksize == erasesize);
             }
 
@@ -246,8 +235,8 @@ static int nxffs_badblocks(FAR struct nxffs_volume_s *volume)
                              volume->pack);
           if (nxfrd != volume->blkper)
             {
-              ferr("ERROR: Write erase block %d failed: %d\n",
-                   lblock, nxfrd);
+              ferr("ERROR: Write erase block %jd failed: %zd\n",
+                   (intmax_t)lblock, nxfrd);
               return -EIO;
             }
         }
